+ Util classes and adaptors: ObservableAsyncList, AsyncTaskService
ObservableAsyncList defines a list in the observer pattern;

AsyncTaskService is a relatively simple asynchronous task runner.
It takes success/error callbacks and also provides a simple applicative functor to run
the callbacks inside a context. This allows for instance running such callbacks in the
UI thread by borrowing the runOnUiThread Runnable consummer.

Signed-off-by: Roosembert Palacios <roosembert.palacios@epfl.ch>